refactor: CSV 파일 읽기 기능 객체 분리
